@@ -3,17 +3,8 @@ description: >-
   Annotation library used to simplify localization
   in Flutter. The library also uses a generator
   dependency called "rosetta_generator".
-version: 0.2.2
-authors:
-- Adam Hosszu <edemlonger@gmail.com>
-- Renato Kiss <renato.kiss.dev@gmail.com>
-- Tamas Agocs <agta91@gmail.com>
+version: 0.3.0-nullsafety.0
 homepage: https://github.com/TeamWanari/rosetta
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
-
-#
-# dev_dependencies:
-#  flutter:
-#    sdk: flutter
+  sdk: ">=2.12.0-0 <3.0.0"
