@@ -6,7 +6,6 @@ class Stone {
   /// Translation files' directory path. The translation files
   /// should be named according to the [Locale.languageCode] of
   /// the represented [Locale].
-  /// Must not be [null].
   final String path;
 
   /// Optional. Indicates which package bundles the referenced
@@ -17,7 +16,7 @@ class Stone {
   /// than the root package, you must declare the resource's host
   /// package or the Localization library will try to load your
   /// JSON assets from the root bundle.
-  final String package;
+  final String? package;
 
   /// Optional. Describes how the keys should be split up to create
   /// sub-keys. The configuration is done using the [Grouping] class.
@@ -29,13 +28,13 @@ class Stone {
   ///
   /// The default [Grouping.separator] String is '.' but
   /// different ones can be declared.
-  final Grouping grouping;
+  final Grouping? grouping;
 
   /// Create an annotation that will generate the Helper and
   /// Delegate classes for the translations located at [path].
   const Stone({
-    this.path,
+    required this.path,
     this.package,
     this.grouping,
-  }) : assert(path != null);
+  });
 }
